Convert StatusActivity into StatusFragment
