+++ app/src/main/java/com/fknussel/yamba/StatusActivity.java
@@ -1,111 +1,27 @@
 package com.fknussel.yamba;
 
-import android.graphics.Color;
-import android.os.AsyncTask;
-import android.support.v7.app.ActionBarActivity;
+import android.app.Activity;
 import android.os.Bundle;
-import android.text.Editable;
-import android.text.TextWatcher;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.marakana.android.yamba.clientlib.YambaClient;
-import com.marakana.android.yamba.clientlib.YambaClientException;
 
 
-public class StatusActivity extends ActionBarActivity {
-
-    private final String TAG = this.getClass().getSimpleName();
-    
-    private Button buttonTweet;
-    private EditText editStatus;
-    private TextView textCount;
-    private int defaultTextColor;
+public class StatusActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_status);
-        
-        buttonTweet = (Button) findViewById(R.id.buttonTweet);
-        editStatus = (EditText) findViewById(R.id.editStatus);
-        textCount = (TextView) findViewById(R.id.textCount);
-        
-        // Clicking on the Tweet button posts the status to the web service
-        buttonTweet.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String status = editStatus.getText().toString();
-                Log.d(TAG, "You just clicked on the TWEET button...");
-                Log.d(TAG, "Your status: " + status);
-                new PostTask().execute(status);
-            }
-        });
-        
-        // Characters remaining
-        defaultTextColor = textCount.getTextColors().getDefaultColor();
-        
-        editStatus.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                
-            }
 
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                // Determine how many characters are left
-                int count = 140 - editStatus.length();
-                
-                // Update the value on screen
-                textCount.setText(Integer.toString(count));
-                
-                // Update the text color accordingly
-                if (count < 10) {
-                    textCount.setTextColor(Color.RED);
-                } else if (count > 10 && count < 50) {
-                    textCount.setTextColor(Color.YELLOW);
-                } else if (count >= 50 && count < 140) {
-                    textCount.setTextColor(Color.GREEN);
-                } else if (count == 140) {
-                    textCount.setTextColor(defaultTextColor);
-                }
-            }
-        });
-    }
-    
-    private final class PostTask extends AsyncTask<String, Void, String> {
-        @Override
-        protected String doInBackground(String... params) {
-            YambaClient yambaCloud = new YambaClient("student", "password");
-            
-            try {
-                yambaCloud.postStatus( params[0] );
-                return "Successfully posted";
-            } catch(YambaClientException e) {
-                e.printStackTrace();
-                return "Failed to post to yamba service";
-            }
-            
-        }
-        
-        @Override
-        protected void onPostExecute(String result) {
-            super.onPostExecute(result);
-            Toast.makeText(StatusActivity.this, result, Toast.LENGTH_SHORT).show();
+        // Check if this activity was created before
+        if (savedInstanceState == null) {
+            // Create a fragment
+            StatusFragment fragment = new StatusFragment();
+            getFragmentManager()
+                    .beginTransaction()
+                    .add(android.R.id.content, fragment, fragment.getClass().getSimpleName())
+                    .commit();
         }
     }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
