+++ app/src/main/res/layout-land/fragment_status.xml
